requirements.txt
# Core
# python>=3.11
streamlit==1.40.0
plotly==5.24.1
pandas==2.2.3
numpy==1.26.4

# Database
PyMySQL==1.1.1
SQLAlchemy==2.0.36
aiomysql==0.2.0
dbutils==3.1.0

# MQTT & Async
paho-mqtt==2.1.0
asyncio-mqtt==0.16.2
aiofiles==24.1.0

# Geospatial
pyproj==3.7.0
shapely==2.0.6
geopandas==1.0.1
rtree==1.3.0
h3==3.7.7
geopy==2.4.1

# Visualization
datashader==0.16.3
pydeck==0.9.1
altair==5.4.1
matplotlib==3.9.2
seaborn==0.13.2

# API
fastapi==0.115.5
uvicorn==0.32.1
websockets==13.1
python-multipart==0.0.12

# Analytics
scikit-learn==1.5.2
scipy==1.14.1
statsmodels==0.14.4

# Performance
redis==5.2.0
cachetools==5.5.0
diskcache==5.6.3

# Monitoring
prometheus-client==0.21.0
psutil==6.1.0

# Utilities
python-dotenv==1.0.1
ruamel.yaml==0.18.6
click==8.1.7
rich==13.9.4
loguru==0.7.2
watchdog==5.0.3

cryptography
networkx